+++ Adj/Algebra.hs
@@ -1,3 +1,4 @@
 module Adj.Algebra (module Exports) where
 
 import Adj.Algebra.Category as Exports
+import Adj.Algebra.Set as Exports