+++ Adj/Algebra/Set.hs
@@ -1,0 +1,1 @@
+module Adj.Algebra.Set where