+++ Adj/Algebra/Set.hs
@@ -23,3 +23,12 @@ absurd x = case x of {}
 type family Neutral p = r | r -> p where
 	Neutral (:*:) = Unit
 	Neutral (:+:) = Void
+
+class Setoid o where
+	(==) :: o -> o -> Unit :+: Unit
+
+	(/=) :: o -> o -> Unit :+: Unit
+	x /= y = case x == y of
+		This _ -> That Unit
+		That _ -> This Unit
+