+++ CHANGELOG.md
@@ -184,3 +184,4 @@
 * Move `Adj.Program.Progress` to `Adj.Program.Primitive` module
 * Move `Adj.Program.Generation` to `Adj.Program.Primitive` module
 * Define `Adj.Program.Primitive.Boolean` module
+* Define `Adj.Algebra.Set` module