+++ CHANGELOG.md
@@ -186,3 +186,5 @@
 * Define `Adj.Program.Primitive.Boolean` module
 * Define `Adj.Algebra.Set` module
 * Move `:*:`, `:+:` to `Adj.Algebra.Set` module
+* Rename `Unit` type family to `Neutral`
+* Reame `Initial` datatype to `Void`