+++ CHANGELOG.md
@@ -194,3 +194,4 @@
 * Move `Void` to `Adj.Algebra.Set` module
 * Move `absurd` to `Adj.Algebra.Set` module
 * Move `boring` to `Adj.Algebra.Set` module
+* Define `Setoid` typeclass