+++ CHANGELOG.md
@@ -205,3 +205,5 @@
 * Define `=:+:=` type alias
 * Define `FGG` scheme
 * Define `Adj.Program.Datastructure.Implementation.Tree` module
+
+# 0.1.0